Dashboard v2: Plugins - split routes, add sub‑nav/menu, default to /plugins/manage
