+++ client/dashboard/app/primary-menu/index.tsx
@@ -11,6 +11,7 @@ function PrimaryMenu() {
 			{ supports.sites && <Menu.Item to="/sites">{ __( 'Sites' ) }</Menu.Item> }
 			{ supports.domains && <Menu.Item to="/domains">{ __( 'Domains' ) }</Menu.Item> }
 			{ supports.emails && <Menu.Item to="/emails">{ __( 'Emails' ) }</Menu.Item> }
+			{ supports.plugins && <Menu.Item to="/plugins">{ __( 'Plugins' ) }</Menu.Item> }
 		</Menu>
 	);
 }
