+++ client/dashboard/app/router/plugins.tsx
@@ -1,5 +1,6 @@
-import { createRoute, createLazyRoute } from '@tanstack/react-router';
+import { createRoute, createLazyRoute, redirect } from '@tanstack/react-router';
 import { rootRoute } from './root';
+import type { AnyRoute } from '@tanstack/react-router';
 
 export const pluginsRoute = createRoute( {
 	getParentRoute: () => rootRoute,
@@ -12,6 +13,37 @@ export const pluginsRoute = createRoute( {
 	)
 );
 
+export const pluginsIndexRoute = createRoute( {
+	getParentRoute: () => pluginsRoute,
+	path: '/',
+	beforeLoad: () => {
+		throw redirect( { to: '/plugins/manage' } );
+	},
+} );
+
+export const pluginsManageRoute = createRoute( {
+	getParentRoute: () => pluginsRoute,
+	path: 'manage',
+} ).lazy( () =>
+	import( '../../plugins/manage' ).then( ( d ) =>
+		createLazyRoute( 'plugins-manage' )( {
+			component: d.default,
+		} )
+	)
+);
+
+export const pluginsScheduleRoute = createRoute( {
+	getParentRoute: () => pluginsRoute,
+	path: 'schedule',
+} ).lazy( () =>
+	import( '../../plugins/schedule' ).then( ( d ) =>
+		createLazyRoute( 'plugins-schedule' )( {
+			component: d.default,
+		} )
+	)
+);
+
 export const createPluginsRoutes = () => {
-	return [ pluginsRoute ];
+	const childRoutes: AnyRoute[] = [ pluginsIndexRoute, pluginsManageRoute, pluginsScheduleRoute ];
+	return [ pluginsRoute.addChildren( childRoutes ) ];
 };
