+++ client/dashboard/plugins/schedule/index.tsx
@@ -1,0 +1,11 @@
+import { __ } from '@wordpress/i18n';
+import { PageHeader } from '../../components/page-header';
+import PageLayout from '../../components/page-layout';
+
+export default function PluginsSchedule() {
+	return (
+		<PageLayout size="small" header={ <PageHeader title={ __( 'Scheduled updates' ) } /> }>
+			{ /* TODO: Implement schedule updates UI */ }
+		</PageLayout>
+	);
+}
